+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.0.2 - 2023-11-02
+
 ### Fixed
 
 * Correction des paramètres d'obtention de la liste des dossiers